+++ Plutarch/Internal/Evaluate.hs
@@ -1,6 +1,15 @@
 {-# OPTIONS_GHC -Wno-incomplete-uni-patterns #-}
 
-module Plutarch.Internal.Evaluate (uplcVersion, evalScript, evalScriptHuge, evalScript', EvalError) where
+module Plutarch.Internal.Evaluate (
+  uplcVersion,
+  evalScript,
+  evalScriptHuge,
+  evalScript',
+  evalTermNoEmit,
+  evalScriptNoEmit,
+  EvalError,
+  NoEmitEvaluateError (..),
+) where
 
 import Data.Text (Text)
 import Plutarch.Script (Script (Script))
@@ -12,12 +21,14 @@ import PlutusCore.Evaluation.Machine.ExBudget (
  )
 import PlutusCore.Evaluation.Machine.ExBudgetingDefaults (defaultCekParameters)
 import PlutusCore.Evaluation.Machine.ExMemory (ExCPU (ExCPU), ExMemory (ExMemory))
+import PlutusCore.Quote qualified as Quote
 import UntypedPlutusCore (
   Program (Program),
   Term,
   Version (Version),
  )
 import UntypedPlutusCore qualified as UPLC
+import UntypedPlutusCore.DeBruijn qualified as DeBruijn
 import UntypedPlutusCore.Evaluation.Machine.Cek qualified as Cek
 
 type EvalError = (Cek.CekEvaluationException PLC.NamedDeBruijn PLC.DefaultUni PLC.DefaultFun)
@@ -56,3 +67,27 @@ evalTerm ::
 evalTerm budget t =
   case Cek.runCekDeBruijn defaultCekParameters (Cek.restricting (ExRestrictingBudget budget)) Cek.logEmitter t of
     (errOrRes, Cek.RestrictingSt (ExRestrictingBudget final), logs) -> (errOrRes, budget `minusExBudget` final, logs)
+
+data NoEmitEvaluateError
+  = NoEmitEvaluateFreeVariableError DeBruijn.FreeVariableError
+  | NoEmitEvaluateFailure
+
+-- Evaluates given script using `unsafeEvaluateCekNoEmit`. This will run with unbound budget and without log output
+evalTermNoEmit ::
+  Term PLC.NamedDeBruijn PLC.DefaultUni PLC.DefaultFun () ->
+  Either NoEmitEvaluateError (Term PLC.NamedDeBruijn PLC.DefaultUni PLC.DefaultFun ())
+evalTermNoEmit t =
+  case Quote.runQuoteT $ DeBruijn.unDeBruijnTerm t of
+    Right namedUPLC ->
+      case Cek.unsafeEvaluateCekNoEmit defaultCekParameters namedUPLC of
+        Cek.EvaluationSuccess evaluated ->
+          case DeBruijn.deBruijnTerm evaluated of
+            Right evaluatedTerm -> Right evaluatedTerm
+            Left err -> Left $ NoEmitEvaluateFreeVariableError err
+        Cek.EvaluationFailure -> Left NoEmitEvaluateFailure
+    Left err -> Left $ NoEmitEvaluateFreeVariableError err
+
+evalScriptNoEmit :: Script -> Either NoEmitEvaluateError Script
+evalScriptNoEmit (Script (Program _ _ script)) =
+  Script . Program () uplcVersion . UPLC.termMapNames UPLC.unNameDeBruijn
+    <$> evalTermNoEmit (UPLC.termMapNames UPLC.fakeNameDeBruijn script)